+++ noodles-fasta/src/async.rs
@@ -1,3 +1,6 @@
-pub mod reader;
+//! Async FASTA.
 
-pub use self::reader::Reader;
+pub mod io;
+
+#[deprecated(since = "0.39.0", note = "Use `gff::r#async::io::Reader` instead.")]
+pub use self::io::Reader;