+++ noodles-fasta/src/async/io.rs
@@ -1,0 +1,5 @@
+//! Async FASTA I/O.
+
+pub(crate) mod reader;
+
+pub use self::reader::Reader;