+++ noodles-util/examples/util_alignment_depth.rs
@@ -22,7 +22,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut builder = alignment::io::indexed_reader::Builder::default();
 
     if let Some(fasta_src) = fasta_src {
-        let repository = fasta::indexed_reader::Builder::default()
+        let repository = fasta::io::indexed_reader::Builder::default()
             .build_from_path(fasta_src)
             .map(fasta::repository::adapters::IndexedReader::new)
             .map(fasta::Repository::new)?;