+++ noodles-util/examples/util_alignment_rewrite.rs
@@ -15,7 +15,7 @@ fn main() -> io::Result<()> {
     let fasta_src = args.next();
 
     let repository = fasta_src
-        .map(|src| fasta::indexed_reader::Builder::default().build_from_path(src))
+        .map(|src| fasta::io::indexed_reader::Builder::default().build_from_path(src))
         .transpose()?
         .map(IndexedReader::new)
         .map(fasta::Repository::new)