+++ noodles-fasta/src/fai/writer.rs
@@ -7,6 +7,47 @@ pub struct Writer<W> {
     inner: W,
 }
 
+impl<W> Writer<W> {
+    /// Returns a reference to the underlying writer.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_fasta::fai;
+    /// let mut writer = fai::Writer::new(Vec::new());
+    /// assert!(writer.get_ref().is_empty());
+    /// ```
+    pub fn get_ref(&self) -> &W {
+        &self.inner
+    }
+
+    /// Returns a mutable reference to the underlying writer.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_fasta::fai;
+    /// let mut writer = fai::Writer::new(Vec::new());
+    /// assert!(writer.get_mut().is_empty());
+    /// ```
+    pub fn get_mut(&mut self) -> &mut W {
+        &mut self.inner
+    }
+
+    /// Returns the underlying writer.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_fasta::fai;
+    /// let mut writer = fai::Writer::new(Vec::new());
+    /// assert!(writer.into_inner().is_empty());
+    /// ```
+    pub fn into_inner(self) -> W {
+        self.inner
+    }
+}
+
 impl<W> Writer<W>
 where
     W: Write,
@@ -21,19 +62,6 @@ where
     /// ```
     pub fn new(inner: W) -> Self {
         Self { inner }
-    }
-
-    /// Returns a reference to the underlying writer.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use noodles_fasta::fai;
-    /// let mut writer = fai::Writer::new(Vec::new());
-    /// assert!(writer.get_ref().is_empty());
-    /// ```
-    pub fn get_ref(&self) -> &W {
-        &self.inner
     }
 
     /// Writes a FASTA index.