+++ noodles-fasta/src/fai/record.rs
@@ -8,7 +8,7 @@ const FIELD_DELIMITER: char = '\t';
 const MAX_FIELDS: usize = 5;
 
 /// A FASTA index record.
-#[derive(Debug, Default, Eq, PartialEq)]
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
 pub struct Record {
     name: Vec<u8>,
     length: u64,